Done with the tweaks and shit. -Gurgled items no longer drop their invisible pocket space holder thingies while dumping their contents. -IDs and probably some other special items shouldn't randomly vanish now. -Transfering contents from gut to another now also transfers their preserved status and cleans up the list entries from source gut. -Ejection, single and complete, now properly clears the lists from ejected stuff. -Some other cleanups yadda yadda. -Added a commented example part hinting that the system is pretty much ready for a trait based item-weak gurgle option. -All should be well now unless you try to gurgle a synth expecting a preserved brain on borgo gut right now. I just remembered I forgot to steal the posibrain yankage bits from the voreorgan code. Oh well it's not like you could've done that with the old code either :v

+++ code/modules/vore/eating/bellymodes_vr.dm
@@ -77,18 +77,28 @@
 				else
 					owner.nutrition += (10/difference)
 
+		//if(owner.weakgurgles < 1) // An example for possible trait modifier. (Not existing currently)
+			//The "Handle leftovers" spaghetti below should be moved here if above gets done.
+
 		// Handle leftovers.
 		var/obj/item/T = pick(internal_contents)
 		if(istype(T, /obj/item))
 			if(istype(T, /obj/item) && _is_digestable(T) && !(T in items_preserved))
-				for(var/obj/item/SubItem in T.contents)
-					SubItem.forceMove(internal_contents)
+				if(T in items_preserved)// Doublecheck just in case.
+					return
+				for(var/obj/item/SubItem in T)
+					if(istype(SubItem,/obj/item/weapon/storage/internal))
+						var/obj/item/weapon/storage/internal/SI = SubItem
+						for(var/obj/item/SubSubItem in SI)
+							SubSubItem.forceMove(internal_contents)
+						qdel(SI)
+					else
+						SubItem.forceMove(internal_contents)
 				owner.nutrition += (1 * T.w_class)
 				internal_contents -= T
 				qdel(T)
 			else
 				return
-
 		return
 
 //////////////////////////// DM_ABSORB ////////////////////////////
